feat: Some optimization

+++ Runtime/Scripts/Injector/Injector.cs
@@ -18,37 +18,37 @@ namespace Shell.Protector
         protected uint rounds = 25;
 
         protected string shader_code_nofilter_XXTEA = @"
-				float4 mip_texture = _MipTex.Sample(sampler_MipTex, mainUV);
+				half4 mip_texture = _MipTex.Sample(sampler_MipTex, mainUV);
 				
 				int mip = round(mip_texture.r * 255 / 10); //fucking precision problems
 				int m[13] = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 10, 10 }; // max size 4k
 
-				float4 c00 = DecryptTextureXXTEA(mainUV, m[mip]);
+				half4 c00 = DecryptTextureXXTEA(mainUV, m[mip]);
 
-				float4 mainTexture = c00;
+				half4 mainTexture = c00;
         ";
         protected string shader_code_bilinear_XXTEA = @"
-				float4 mip_texture = _MipTex.Sample(sampler_MipTex, mainUV);
+				half4 mip_texture = _MipTex.Sample(sampler_MipTex, mainUV);
 				
-				float2 uv_unit = _MainTex_TexelSize.xy;
+				half2 uv_unit = _MainTex_TexelSize.xy;
 				//bilinear interpolation
-				float2 uv_bilinear = poiMesh.uv[0] - 0.5 * uv_unit;
+				half2 uv_bilinear = poiMesh.uv[0] - 0.5 * uv_unit;
 				int mip = round(mip_texture.r * 255 / 10); //fucking precision problems
 				int m[13] = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 10, 10 }; // max size 4k
 				
-                float4 c00 = DecryptTextureXXTEA(uv_bilinear + float2(uv_unit.x * 0, uv_unit.y * 0), m[mip]);
-                float4 c10 = DecryptTextureXXTEA(uv_bilinear + float2(uv_unit.x * 1, uv_unit.y * 0), m[mip]);
-                float4 c01 = DecryptTextureXXTEA(uv_bilinear + float2(uv_unit.x * 0, uv_unit.y * 1), m[mip]);
-                float4 c11 = DecryptTextureXXTEA(uv_bilinear + float2(uv_unit.x * 1, uv_unit.y * 1), m[mip]);
+                half4 c00 = DecryptTextureXXTEA(uv_bilinear + half2(uv_unit.x * 0, uv_unit.y * 0), m[mip]);
+                half4 c10 = DecryptTextureXXTEA(uv_bilinear + half2(uv_unit.x * 1, uv_unit.y * 0), m[mip]);
+                half4 c01 = DecryptTextureXXTEA(uv_bilinear + half2(uv_unit.x * 0, uv_unit.y * 1), m[mip]);
+                half4 c11 = DecryptTextureXXTEA(uv_bilinear + half2(uv_unit.x * 1, uv_unit.y * 1), m[mip]);
 				
-				float2 f = frac(uv_bilinear * _MainTex_TexelSize.zw);
+				half2 f = frac(uv_bilinear * _MainTex_TexelSize.zw);
 				
-				float4 c0 = lerp(c00, c10, f.x);
-				float4 c1 = lerp(c01, c11, f.x);
+				half4 c0 = lerp(c00, c10, f.x);
+				half4 c1 = lerp(c01, c11, f.x);
 
-				float4 bilinear = lerp(c0, c1, f.y);
+				half4 bilinear = lerp(c0, c1, f.y);
 				
-				float4 mainTexture = bilinear;
+				half4 mainTexture = bilinear;
         ";
 
         protected GameObject target;
